ocr_service/requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
pillow
python-multipart